Reorganize conference feature files
