+++ react/features/conference/components/web/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as Conference } from './Conference';
